requirements.txt
pygame-ce
matplotlib